Rename TestPlatform to Runtime
This clarifies the distinction between TestPlatform and SuitePlatform,
and uses a more specific word rather than a more general one.
Unfortunately, it does mean that we have some user-facing
interfaces (the --platform flag, the override_platforms and
define_platforms configuration fields) that now use different
terminology than the test runner.

It would be possible to migrate those to refer to "runtime" as well,
but seeing as --platform in particular is very widely-used *and*
matches up nicely with --preset (-p vs -P) I'm inclined to think it's
not worth the effort.